requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0
PyYAML>=5.4.0
matplotlib>=3.3.0
seaborn>=0.11.0
tqdm>=4.60.0
tensorboard>=2.5.0
scikit-learn>=0.24.0
albumentations>=1.0.0
onnx>=1.10.0
onnxruntime>=1.9.0
pandas>=1.3.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
gradio>=4.0.0
requests>=2.25.0
wandb>=0.15.0